requirements.txt
altair==5.2.0
appdirs==1.4.4
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
frozendict==2.4.0
gitdb==4.0.11
GitPython==3.1.42
html5lib==1.1
idna==3.6
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multitasking==0.0.11
numpy==1.26.4
packaging==23.2
pandas==2.2.1
peewee==3.17.1
pillow==10.2.0
plotly==5.19.0
protobuf==4.25.3
pyarrow==15.0.1
pydeck==0.8.1b0
pygments==2.17.2
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.33.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
streamlit==1.32.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
webencodings==0.5.1
yfinance==0.2.37